Making tokenizing methods thread-safe
We pass streambuf as an argument to the different tokenizing methods instead of having it as a class attribute. We can then remove the streambuf `sb` attribute from the class.
We remove the istream `in` attribute and the constructor initializing it as well.
For a bit more clarity, we renamed sb to sbuf.